When copying to clipboard fails, persist anyway
Via issue #8

The previous behavior is that when copying to the clipboard fails during
a rooster commands, the password would be printed but not persisted.
With this change, it will be printed and persisted.

Other possible approaches:

* don't print the password, tell the user the operation failed
* don't print the password, tell the user how they can see it if they
  want to (`rooster get site --show`)

I included `rooster get` in these changes for consistency. I wasn't
concerned about a change not being persisted, but I didn't *think* it
should exit 1 if all of the other instances of clipboard error handling
treat it as an acceptable circumstance.

The one thing I'm wondering is whether rooster should notify the user
that something went wrong during the copying so they can attempt to fix
it. What do you think, Conrad?